+++ .github/workflows/dotnet.yml
@@ -20,7 +20,7 @@ jobs:
     - name: Restore dependencies UA-CloudLibrary
       run: dotnet restore UACloudLibraryServer\UA-CloudLibrary.csproj
     - name: Restore dependencies UACloudLibClientLibrary
-      run: dotnet restore UACloudLibraryGraphQLClient\UACloudLibClientLibrary\UACloudLibClientLibrary.csproj
+      run: dotnet restore UACloudLibClientLibrary\UACloudLibClientLibrary.csproj
     - name: Restore dependencies SampleConsoleClient
       run: dotnet restore SampleConsoleClient\SampleConsoleClient.csproj
     - name: Build
