+++ UACloudLibClientLibrary/Models/QueryModel.cs
